INS-126: Removed redundent storage implementation.
