+++ ledger/storage/iterate.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package badgerdb
+package storage
 
 import (
 	"bytes"
