+++ ledger/storage/storagetest/storage_test.go
@@ -14,25 +14,62 @@
  *    limitations under the License.
  */
 
-package badgerdb_test
+package storagetest
 
 import (
+	"crypto/rand"
+	"encoding/hex"
 	"fmt"
 	"testing"
 
+	"github.com/insolar/insolar/ledger/storage"
 	"github.com/stretchr/testify/assert"
 
 	"github.com/insolar/insolar/ledger/index"
 	"github.com/insolar/insolar/ledger/jetdrop"
 	"github.com/insolar/insolar/ledger/record"
-	"github.com/insolar/insolar/ledger/storage"
-
-	"github.com/insolar/insolar/ledger/storage/badgerdb/badgertestutils"
 )
+
+func zerohash() []byte {
+	b := make([]byte, record.HashSize)
+	return b
+}
+
+func randhash() []byte {
+	b := zerohash()
+	_, err := rand.Read(b)
+	if err != nil {
+		panic(err)
+	}
+	return b
+}
+
+func hexhash(hash string) []byte {
+	b := zerohash()
+	if len(hash)%2 == 1 {
+		hash = "0" + hash
+	}
+	h, err := hex.DecodeString(hash)
+	if err != nil {
+		panic(err)
+	}
+	_ = copy(b, h)
+	return b
+}
+
+func referenceWithHashes(domainhash, recordhash string) record.Reference {
+	dh := hexhash(domainhash)
+	rh := hexhash(recordhash)
+
+	return record.Reference{
+		Domain: record.ID{Hash: dh},
+		Record: record.ID{Hash: rh},
+	}
+}
 
 func TestStore_GetRecordNotFound(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	ref := &record.Reference{}
@@ -43,7 +80,7 @@ func TestStore_GetRecordNotFound(t *testing.T) {
 
 func TestStore_SetRecord(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 	// mock pulse source
 	pulse1 := record.PulseNum(1)
@@ -80,7 +117,7 @@ func TestStore_SetRecord(t *testing.T) {
 
 func TestStore_GetClassIndex_ReturnsNotFoundIfNoIndex(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	ref := &record.Reference{
@@ -94,7 +131,7 @@ func TestStore_GetClassIndex_ReturnsNotFoundIfNoIndex(t *testing.T) {
 
 func TestStore_SetClassIndex_StoresCorrectDataInStorage(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	zerodomain := record.ID{Hash: zerohash()}
@@ -128,7 +165,7 @@ func TestStore_SetClassIndex_StoresCorrectDataInStorage(t *testing.T) {
 
 func TestStore_SetObjectIndex_ReturnsNotFoundIfNoIndex(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	ref := referenceWithHashes("1000", "5000")
@@ -139,7 +176,7 @@ func TestStore_SetObjectIndex_ReturnsNotFoundIfNoIndex(t *testing.T) {
 
 func TestStore_SetObjectIndex_StoresCorrectDataInStorage(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	idx := index.ObjectLifeline{
@@ -162,7 +199,7 @@ func TestStore_SetObjectIndex_StoresCorrectDataInStorage(t *testing.T) {
 
 func TestStore_GetDrop_ReturnsNotFoundIfNoDrop(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	drop, err := store.GetDrop(1)
@@ -172,7 +209,7 @@ func TestStore_GetDrop_ReturnsNotFoundIfNoDrop(t *testing.T) {
 
 func TestStore_SetDrop_StoresCorrectDataInStorage(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	// it references on 'fake' zero
@@ -190,7 +227,7 @@ func TestStore_SetDrop_StoresCorrectDataInStorage(t *testing.T) {
 
 func TestStore_SetCurrentPulse(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	store.SetCurrentPulse(42)
@@ -199,7 +236,7 @@ func TestStore_SetCurrentPulse(t *testing.T) {
 
 func TestStore_SetEntropy(t *testing.T) {
 	t.Parallel()
-	store, cleaner := badgertestutils.TmpDB(t, "")
+	store, cleaner := TmpStore(t, "")
 	defer cleaner()
 
 	store.SetEntropy(42, []byte{1, 2, 3})
