+++ ledger/storage/transaction.go
@@ -1,4 +1,20 @@
-package badgerdb
+/*
+ *    Copyright 2018 INS Ecosystem
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package storage
 
 import (
 	"github.com/dgraph-io/badger"
@@ -6,32 +22,41 @@ import (
 	"github.com/insolar/insolar/ledger/index"
 	"github.com/insolar/insolar/ledger/jetdrop"
 	"github.com/insolar/insolar/ledger/record"
-	"github.com/insolar/insolar/ledger/storage"
 )
 
+// TransactionManager is used to ensure persistent writes to disk.
 type TransactionManager struct {
 	store *Store
 	txn   *badger.Txn
 }
 
+func prefixkey(prefix byte, key []byte) []byte {
+	k := make([]byte, record.RefIDSize+1)
+	k[0] = prefix
+	_ = copy(k[1:], key)
+	return k
+}
+
+// Commit tries to write transaction on disk. Returns error on fail.
 func (m *TransactionManager) Commit() error {
 	// TODO: check transaction conflict
 	return m.txn.Commit(nil)
 }
 
+// Discard terminates transaction without disk writes.
 func (m *TransactionManager) Discard() {
 	m.txn.Discard()
 }
 
 // GetRecord returns record from BadgerDB by *record.Reference.
 //
-// It returns storage.ErrNotFound if the DB does not contain the key.
+// It returns ErrNotFound if the DB does not contain the key.
 func (m *TransactionManager) GetRecord(ref *record.Reference) (record.Record, error) {
 	k := prefixkey(scopeIDRecord, ref.CoreRef()[:])
 	item, err := m.txn.Get(k)
 	if err != nil {
 		if err == badger.ErrKeyNotFound {
-			return nil, storage.ErrNotFound
+			return nil, ErrNotFound
 		}
 		return nil, err
 	}
@@ -40,6 +65,9 @@ func (m *TransactionManager) GetRecord(ref *record.Reference) (record.Record, er
 		return nil, err
 	}
 	raw, err := record.DecodeToRaw(buf)
+	if err != nil {
+		return nil, err
+	}
 
 	return raw.ToRecord(), nil
 }
@@ -73,7 +101,7 @@ func (m *TransactionManager) GetClassIndex(ref *record.Reference) (*index.ClassL
 	item, err := m.txn.Get(k)
 	if err != nil {
 		if err == badger.ErrKeyNotFound {
-			return nil, storage.ErrNotFound
+			return nil, ErrNotFound
 		}
 		return nil, err
 	}
@@ -100,7 +128,7 @@ func (m *TransactionManager) GetObjectIndex(ref *record.Reference) (*index.Objec
 	item, err := m.txn.Get(k)
 	if err != nil {
 		if err == badger.ErrKeyNotFound {
-			return nil, storage.ErrNotFound
+			return nil, ErrNotFound
 		}
 		return nil, err
 	}
@@ -127,7 +155,7 @@ func (m *TransactionManager) GetDrop(pulse record.PulseNum) (*jetdrop.JetDrop, e
 	item, err := m.txn.Get(k)
 	if err != nil {
 		if err == badger.ErrKeyNotFound {
-			return nil, storage.ErrNotFound
+			return nil, ErrNotFound
 		}
 		return nil, err
 	}
@@ -181,14 +209,6 @@ func (m *TransactionManager) SetDrop(pulse record.PulseNum, prevdrop *jetdrop.Je
 	return drop, err
 }
 
-// SetEntropy stores given entropy for given pulse in storage.
-//
-// Entropy is used for calculating node roles.
-func (m *TransactionManager) SetEntropy(pulse record.PulseNum, entropy []byte) error {
-	k := prefixkey(scopeIDEntropy, record.EncodePulseNum(pulse))
-	return m.txn.Set(k, entropy)
-}
-
 // GetEntropy returns entropy from storage for given pulse.
 //
 // Entropy is used for calculating node roles.
@@ -197,7 +217,7 @@ func (m *TransactionManager) GetEntropy(pulse record.PulseNum) ([]byte, error) {
 	item, err := m.txn.Get(k)
 	if err != nil {
 		if err == badger.ErrKeyNotFound {
-			return nil, storage.ErrNotFound
+			return nil, ErrNotFound
 		}
 		return nil, err
 	}
@@ -206,4 +226,12 @@ func (m *TransactionManager) GetEntropy(pulse record.PulseNum) ([]byte, error) {
 		return nil, err
 	}
 	return buf, nil
+}
+
+// SetEntropy stores given entropy for given pulse in storage.
+//
+// Entropy is used for calculating node roles.
+func (m *TransactionManager) SetEntropy(pulse record.PulseNum, entropy []byte) error {
+	k := prefixkey(scopeIDEntropy, record.EncodePulseNum(pulse))
+	return m.txn.Set(k, entropy)
 }
